requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
azure-common==1.1.28
azure-core==1.26.3
azure-functions==1.13.3
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
backoff==1.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
multidict==6.0.4
pycparser==2.21
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
spotify==0.10.2
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
